requirements.txt
Django==5.1
djangorestframework
pillow
djangorestframework-simplejwt
django-filter
django-taggit==6.0.0
python-dotenv
psycopg2-binary==2.9.9